Added functionality to test by including many more functions
